requirements.txt
opencv-python
numpy
imutils
scikit-image
Pillow
